#9 Fix support of numeric character state names
In the code prior to the block that was changed, we peek for comma and semicolon - either one of those puts the first char_state name, 18, on lexer's @next_token (because it wasn't a semicolon or a comma, but it was a Number token). Now we start the while loop:

the peek value is not a label, it is a number (because it was already peeked and put in lexer's @next_token).

Now the next iteration of the while loop we peek with @input "17 16 ..." - first we peek for a label, which it is ('17'). In the loop we pop the label ('17') and then immediately - still within the while loop - check for and pop a number (16). Now we pass to the next iteration and repeat.

If you only have two numbers in a row, the first is parsed as a Number 18 (before we even get to the while loop), the second is parsed as a label '17' in the second iteration of the while loop, and then there's not a third number to be parsed as a Number, so all is fine.

If you have 3 tokens in a row that are each both Number and Label (really the first token can be a Number that's not a Label) then the first iteration gets a Number 18 and adds it to the `opts` hash with key index 0, the second adds an `opt` on key index 1 for Label '17', then within the same loop adds key index 1 for Number 16, shadowing the '17'. The following iterations repeat the same process of shadowing every other number.

Note that neither of Label and Number contains the other, so we do want to check for both:
* -3 is a Number but not a Label
* 'asdf' is a Label but not a Number