#16 Only match the requested Peek Token in Lexer
The main issue this fixes is overlap of Tokens when combined with Lexer's peek caching.

The specific issue here was trying to tokenize character state names. Say we're trying to parse `123asdf` as a character state name. Our first step is to peek on the lexer for a semicolon to see if we're already at the end. In this case that fails but the peek stores `Tokens::Number: 123` on the lexer as the next token (Number is tried before Label is). Now we want to allow Numbers or Labels as char state name tokens (they're different classes), so we either have to peek for a Number or a Label at this point.

If we peek for a Number we get `123` but now what's left of the name is `asdf` and the lexer doesn't expose if it was originally the string `123 abc` - in which case we should parse two separate tokens - or if it was really `123abc` in which case we could read a Label again to get the whole name. No way to know without exposing more information from the lexer.

If we instead peek for a Label first then we fail because what's on the lexer already is a Number token, not a Label token. At that point we're out of luck.

It's even worse though: what if my character state name was `a=b`? Then peeking for a semicolon puts a ValuePair on the lexer and now peeking for neither Number nor Label will find anything. Similarly with `ids...` as both a MesquiteId and a Label, etc. etc. Many things overlap with Label.

At this point one might object that now we'll never know when we're trying to peek an unrecognized token, but I think that was actually always the case.

Note that RowVec matches *every* non-nil @input (since every token slurps whitespace).

That means that we were *never* erroring out on peek matching no token - we were erroring out because instead of any of the tokens we thought we might see in a particular block, we wound up matching something like RowVec we didn't expect to see and then wound up in an infinite loop because none of our peeks worked anymore because a RowVec was on the lexer when nobody was peeking for a RowVec (these are the "gazillion ..." errors).

Removing the 'no token recognized' error just accepts that fact in a different way: we still error out on the gazillion/infinite loop - still because we're not seeing any of the tokens we expected to see - but now we're just storing nothing on the lexer when we don't find a match instead of storing something broad like RowVec that was bogus anyway.

If you *want* to match RowVec you need to peek for it or just pop and pray.

The Parser::parse_notes_blk had a couple builtin assumptions about peek holding the next token it matched if it didn't match the token that was peeked for:

it was checking for a Label before it checked for FileLbl - that used to work because FileLbl was above Label in the token search order: now you have to explicitly search for the more restrictive token first if that's what you want (it's no longer implicit in the order of the tokens search list).

it used to put a Label on the lexer with a peek and then rely on that token staying on the lexer through multiple other peeks on the next pass through the while loop until it got back to where Labels were being processed again. It didn't need to work that way (if you did need something like that you'd need to use a local variable instead).